+++ include/xal/Source.h
@@ -35,7 +35,7 @@ namespace xal
 		
 		unsigned int getSourceId() { return this->sourceId; }
 		void setSourceId(unsigned int value) { this->sourceId = value; }
-		float getSampleOffset() { return this->sampleOffset; }
+		int getSampleOffset() { return this->sampleOffset; }
 		unsigned int getBuffer();
 		SoundBuffer* getSound() { return this->sound; }
 		bool isBound() { return this->bound; }
@@ -57,7 +57,7 @@ namespace xal
 		bool paused;
 		float fadeSpeed;
 		float fadeTime;
-		float sampleOffset;
+		int sampleOffset;
 		bool bound;
 		SoundBuffer* sound;
